feat: get observing transaction index
